Move EuiSuperDatePicker harness to test-eui-helpers
